+++ evalml/automl/automl_search.py
@@ -289,8 +289,11 @@ class AutoMLSearch:
         self.pipeline_parameters = pipeline_parameters if pipeline_parameters is not None else {}
         self.search_iteration_plot = None
         self._interrupted = False
+        self.frozen_pipeline_parameters = {}
+
         if len(self.problem_configuration):
             pipeline_params = {**{'pipeline': self.problem_configuration}, **self.pipeline_parameters}
+            self.frozen_pipeline_parameters = {'pipeline': self.problem_configuration}
         else:
             pipeline_params = self.pipeline_parameters
 
@@ -299,13 +302,14 @@ class AutoMLSearch:
             allowed_estimators = get_estimators(self.problem_type, self.allowed_model_families)
             logger.debug(f"allowed_estimators set to {[estimator.name for estimator in allowed_estimators]}")
             drop_columns = self.pipeline_parameters['Drop Columns Transformer']['columns'] if 'Drop Columns Transformer' in self.pipeline_parameters else None
-            if len(index_columns) > 0 and drop_columns is None:
-                self.pipeline_parameters['Drop Columns Transformer'] = {'columns': index_columns}
-            if is_time_series(self.problem_type):
-                self.allowed_pipelines = [make_pipeline(self.X_train, self.y_train, estimator, self.problem_type, pipeline_params, custom_hyperparameters=pipeline_params) for estimator in allowed_estimators]
-            else:
-                self.allowed_pipelines = [make_pipeline(self.X_train, self.y_train, estimator, self.problem_type, None, custom_hyperparameters=pipeline_params) for estimator in allowed_estimators]
             index_columns = list(self.X_train.select('index').columns)
+            if len(index_columns) > 0 and drop_columns is None:
+                self.frozen_pipeline_parameters['Drop Columns Transformer'] = {'columns': index_columns}
+
+            if is_time_series(self.problem_type):
+                self.allowed_pipelines = [make_pipeline(self.X_train, self.y_train, estimator, self.problem_type, self.frozen_pipeline_parameters, custom_hyperparameters=pipeline_params) for estimator in allowed_estimators]
+            else:
+                self.allowed_pipelines = [make_pipeline(self.X_train, self.y_train, estimator, self.problem_type, self.frozen_pipeline_parameters, custom_hyperparameters=pipeline_params) for estimator in allowed_estimators]
         else:
             for pipeline in self.allowed_pipelines:
                 if self.pipeline_parameters:
@@ -380,7 +384,8 @@ class AutoMLSearch:
             number_features=self.X_train.shape[1],
             pipelines_per_batch=self._pipelines_per_batch,
             ensembling=run_ensembling,
-            pipeline_params=pipeline_params
+            pipeline_params=pipeline_params,
+            frozen_pipeline_parameters=self.frozen_pipeline_parameters
         )
 
     def _get_batch_number(self):
